requirements.txt
# DAQ Pressure Sensor Monitor Dependencies (Ubuntu/Linux)
# Install with: pip install -r requirements.txt

# GUI Framework
PyQt5>=5.15.0
pyqtgraph>=0.12.0
numpy>=1.21.0
matplotlib>=3.5.0

# MCC DAQ Hardware Support (Primary)
uldaq>=1.2.0

# Fallback DAQ Support
pyusb>=1.2.1
pyserial>=3.5
hidapi>=0.12.0

# Optional: For better USB device information
# libusb1>=1.9.3